backend/requirements.txt
fastapi
uvicorn
pillow
torch
torchvision
ultralytics
firebase-admin
faiss-cpu
python-multipart
